#382 Integration tests should run with RMQ and HTTP messaging
