+++ miscellaneous/roboconf-integration-tests/src/test/java/net/roboconf/integration/tests/paxrunner/messaging/AgentInMemoryWithHttpTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2014-2016 Linagora, Université Joseph Fourier, Floralis
+ * Copyright 2016 Linagora, Université Joseph Fourier, Floralis
  *
  * The present code is developed in the scope of the joint LINAGORA -
  * Université Joseph Fourier - Floralis research program and is designated
@@ -23,23 +23,20 @@
  * limitations under the License.
  */
 
-package net.roboconf.integration.probes;
+package net.roboconf.integration.tests.paxrunner.messaging;
 
-import net.roboconf.integration.tests.internal.ItUtils;
-import net.roboconf.integration.tests.internal.parameterized.RabbitMqConfiguration;
+import net.roboconf.integration.tests.internal.parameterized.HttpConfiguration;
 
-import org.ops4j.pax.exam.Configuration;
-import org.ops4j.pax.exam.Option;
 
 /**
- * A base class to run the DM's distribution.
  * @author Vincent Zurczak - Linagora
  */
-public abstract class DmTest {
+public class AgentInMemoryWithHttpTest extends AbstractAgentInMemoryTest {
 
-	@Configuration
-	public Option[] config() throws Exception {
-		ItConfigurationBean bean = new ItConfigurationBean( "roboconf-karaf-dist-dm", "dm" );
-		return ItUtils.getBaseOptions( bean, new RabbitMqConfiguration());
+	/**
+	 * Constructor.
+	 */
+	public AgentInMemoryWithHttpTest() {
+		super( new HttpConfiguration(), "HTTP" );
 	}
 }
