+++ miscellaneous/roboconf-integration-tests/src/test/java/net/roboconf/integration/tests/paxrunner/messaging/AgentInMemoryWithRabbitMqTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2015-2016 Linagora, Université Joseph Fourier, Floralis
+ * Copyright 2016 Linagora, Université Joseph Fourier, Floralis
  *
  * The present code is developed in the scope of the joint LINAGORA -
  * Université Joseph Fourier - Floralis research program and is designated
@@ -23,16 +23,19 @@
  * limitations under the License.
  */
 
-package net.roboconf.integration.tests.internal.parametrized;
+package net.roboconf.integration.tests.paxrunner.messaging;
 
-import java.util.List;
-
-import org.ops4j.pax.exam.Option;
+import net.roboconf.integration.tests.internal.parameterized.RabbitMqConfiguration;
 
 /**
  * @author Vincent Zurczak - Linagora
  */
-public interface IMessagingConfiguration {
+public class AgentInMemoryWithRabbitMqTest extends AbstractAgentInMemoryTest {
 
-	List<Option> options();
+	/**
+	 * Constructor.
+	 */
+	public AgentInMemoryWithRabbitMqTest() {
+		super( new RabbitMqConfiguration(), "Rabbit MQ" );
+	}
 }
